+++ src/main/scala/pro/civitaspo/digdag/plugin/ecs_task/command/EcsTaskCommandRunner.scala
@@ -9,7 +9,14 @@ import pro.civitaspo.digdag.plugin.ecs_task.aws.AwsConf
 
 import scala.collection.JavaConverters._
 
-case class EcsTaskCommandRunner(params: Config, environments: Map[String, String], awsConf: AwsConf, logger: Logger) {
+case class EcsTaskCommandRunner(
+  scriptsLocationPrefix: AmazonS3URI,
+  script: String,
+  params: Config,
+  environments: Map[String, String],
+  awsConf: AwsConf,
+  logger: Logger
+) {
 
   val cf: ConfigFactory = params.getFactory
 
@@ -20,15 +27,11 @@ case class EcsTaskCommandRunner(params: Config, environments: Map[String, String
   val cpu: Optional[String] = params.getOptional("cpu", classOf[String])
   val executionRoleArn: Optional[String] = params.getOptional("execution_role_arn", classOf[String])
 
-  val family: String = params.get(
-    "family",
-    classOf[String],
-    params
-      .get("task_name", classOf[String])
-      .replaceAll("\\+", "_")
-      .replaceAll("\\^", "_")
-      .replaceAll("\\=", "_")
-  )
+  val taskName: String = params.get("task_name", classOf[String])
+  val familyPrefix: String = params.get("family_prefix", classOf[String], "")
+  val familySuffix: String = params.get("family_suffix", classOf[String], "")
+  val familyInfix: String = params.get("family_infix", classOf[String], taskName.replaceAll("\\+", "_").replaceAll("\\^", "_").replaceAll("\\=", "_"))
+  val family: String = params.get("family", classOf[String], s"$familyPrefix$familyInfix$familySuffix")
   val memory: Optional[String] = params.getOptional("memory", classOf[String])
   val networkMode: Optional[String] = params.getOptional("network_mode", classOf[String])
   // NOTE: Use `ecs_task.run>`'s one.
@@ -95,22 +98,22 @@ case class EcsTaskCommandRunner(params: Config, environments: Map[String, String
   // For ecs_task.wait operator
   val timeout: DurationParam = params.get("timeout", classOf[DurationParam], DurationParam.parse("15m"))
 
-  def run(scriptsLocationPrefix: AmazonS3URI): TaskResult = {
+  def run(): TaskResult = {
     val subTasks: Config = cf.create()
-    subTasks.setNested("+register", ecsTaskRegisterSubTask(scriptsLocationPrefix))
+    subTasks.setNested("+register", ecsTaskRegisterSubTask())
     subTasks.setNested("+run", ecsTaskRunInternalSubTask())
     subTasks.setNested("+wait", ecsTaskWaitSubTask())
-    subTasks.setNested("+result", ecsTaskResultSubTask(scriptsLocationPrefix))
+    subTasks.setNested("+result", ecsTaskResultSubTask())
 
     val builder = TaskResult.defaultBuilder(cf)
     builder.subtaskConfig(subTasks)
     builder.build()
   }
 
-  protected def ecsTaskRegisterSubTask(scriptsLocationPrefix: AmazonS3URI): Config = {
+  protected def ecsTaskRegisterSubTask(): Config = {
     withDefaultSubTask { subTask =>
       subTask.set("_type", "ecs_task.register")
-      subTask.set("_command", taskDefinitionConfig(scriptsLocationPrefix))
+      subTask.set("_command", taskDefinitionConfig())
     }
   }
 
@@ -141,10 +144,10 @@ case class EcsTaskCommandRunner(params: Config, environments: Map[String, String
     }
   }
 
-  protected def ecsTaskResultSubTask(resultLocationPrefix: AmazonS3URI): Config = {
+  protected def ecsTaskResultSubTask(): Config = {
     withDefaultSubTask { subTask =>
       subTask.set("_type", "ecs_task.command_result_internal")
-      subTask.set("_command", resultLocationPrefix.toString)
+      subTask.set("_command", scriptsLocationPrefix.toString)
     }
   }
 
@@ -162,10 +165,10 @@ case class EcsTaskCommandRunner(params: Config, environments: Map[String, String
     subTask
   }
 
-  protected def taskDefinitionConfig(scriptsLocationPrefix: AmazonS3URI): Config = {
+  protected def taskDefinitionConfig(): Config = {
     val c: Config = cf.create()
 
-    c.set("container_definitions", (Seq(containerDefinitionConfig(scriptsLocationPrefix)) ++ sidecars).asJava)
+    c.set("container_definitions", (Seq(containerDefinitionConfig()) ++ sidecars).asJava)
     c.setOptional("cpu", cpu)
     c.setOptional("execution_role_arn", executionRoleArn)
     c.set("family", family)
@@ -178,16 +181,16 @@ case class EcsTaskCommandRunner(params: Config, environments: Map[String, String
     c
   }
 
-  protected def containerDefinitionConfig(scriptsLocationPrefix: AmazonS3URI): Config = {
+  protected def containerDefinitionConfig(): Config = {
     val c: Config = cf.create()
 
-    val command: Seq[String] = Seq("sh", "-c", s"aws s3 cp ${scriptsLocationPrefix.toString}/run.sh ./ && sh run.sh")
+    val command: Seq[String] = Seq("sh", "-c", s"aws s3 cp ${scriptsLocationPrefix.toString}/$script ./ && sh $script")
     logger.info(s"Run in the container: ${command.mkString(" ")}")
     c.set("command", command.asJava)
     c.setOptional("disable_networking", disableNetworking)
     c.set("dns_search_domains", dnsSearchDomains.asJava)
     c.set("dns_servers", dnsServers.asJava)
-    val additionalLabels: Map[String, String] = Map("pro.civitaspo.digdag.plugin.ecs_task.version" -> "0.0.3")
+    val additionalLabels: Map[String, String] = Map("pro.civitaspo.digdag.plugin.ecs_task.version" -> "0.0.4")
     c.set("docker_labels", (dockerLabels ++ additionalLabels).asJava)
     c.set("entry_point", entryPoint.asJava)
     c.set("environment", (configEnvironment ++ environments).asJava)